requirements.txt
numpy
pandas
plotly.express
streamlit
json
re
functools
datetime
unicodedata